06_ 기능 - RestController 예외처리 (insert 상품명 중복체크 / delete 상품Id 중복체크 / update 상품Id 중복체크) 구현 완료

+++ src/main/java/site/metacoding/firstapp/web/ProductApiController.java
@@ -34,18 +34,40 @@ public class ProductApiController {
 	
 	@PostMapping("/product/insert")
 	public int insert(@RequestBody Product product) {
+		
+		// 1. db에 productName으로 등록된 상품이 있는지 유무 확인
+		Product productPS = productDao.findByName(product.getProductName());
+		if(productPS != null) { // 2. 상품이 있으면 다음 코드 실행
+			throw new RuntimeException("해당 상품 있음");
+		}
+		// 3. 상품이 없으면 다음 코드 실행
 		int productInsert = productDao.insert(product);
 		return productInsert;
 	}
 	
 	@DeleteMapping("/product/{productId}/delete")
 	public int deleteById(@PathVariable Integer productId) {
+		
+		// 1. db에 productId로 등록된 상품이 있는지 유무 확인
+		Product productPS = productDao.findById(productId);
+
+		if(productPS == null) { // 2. 상품이 없으면 다음 코드 실행
+			throw new RuntimeException("해당 상품 없음");
+			
+		} // 3. 상품이 있으면 다음 코드 실행
 		int productDelete = productDao.deleteById(productId);
 		return productDelete;
 	}
 	
 	@PutMapping("/product/edit")
 	public int update(@RequestBody Product product) {
+		
+		// 1. db에 productId로 등록된 상품이 있는지 유무 확인
+		Product productPS = productDao.findById(product.getProductId());
+		if(productPS == null) { // 2. 상품이 없으면 다음 코드 실행
+			throw new RuntimeException("해당 상품 없음");
+		}
+		// 3. 상품이 있으면 다음 코드 실행
 		int productEdit = productDao.update(product);
 		return productEdit;
 	}
